Clean up search algorithm option
Keep the default one which was skim and removed
the ability to pass algorithm to the executable, which wasn't
being used in the extension
