+++ pylight/Cargo.toml
@@ -23,7 +23,6 @@ fluent-uri = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 crossterm = "0.29"
 rustyline = "15.0"
-nucleo-matcher = "0.3"
 lsp-server = "0.7.8"
 lsp-types = "0.97.0"
 serde_json = "1.0.140"
