+++ pylight/src/lib.rs
@@ -2,5 +2,4 @@ pub mod files;
 pub mod path_trie;
 pub mod python;
 pub mod search;
-pub mod search_skim;
 pub mod symbols;
